Simplify the table definition for the unique lookup vindex
The `corder_id` is not an auto generated column in practice, because it
uses the `corder_id` values from the `corder` table here.

Having this an auto increment doesn't break, but since the value is
always provided it's not necessary and having it is confusing (as it
might be thought the value is not explicitly supplied).

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>