Allow empty commits
This is to support the preview functionality of the bot.

The root "cause" here is: on the _first_ PR opened against a release
branch after cutting a release tag from that series, the release tag and
the head of the release branch are the same, so there is no change
generated by `make generated-docs`. This would then cause both the
`git add` and `git commit` to error.

Signed-off-by: Andrew Mason <andrew@planetscale.com>